drop loadMap
