+++ test/render.test.js
@@ -3,14 +3,19 @@
 /* jshint node:true */
 
 var TileSource = require('..');
+var fileSource = require('./lib/fs');
+var GL = TileSource(fileSource);
 var test = require('tape').test;
-var loadMap = require('./lib/load');
 var fs = require('fs');
 var path = require('path');
 var mkdirp = require('mkdirp');
 var compare = require('./lib/compare')
 var tiles = ['0-0-0', '1-0-1', '2-1-1', '3-2-3', '4-4-6'];
 var style = require('./fixtures/style.json');
+
+TileSource.mbgl.on('message', function(msg) {
+    console.log(msg);
+});
 
 function filePath(name) {
     return ['expected', 'actual', 'diff'].reduce(function(prev, key) {
@@ -33,7 +38,7 @@ function render(name, scale) {
     var filename = filePath(name + (scale ? '@' + scale + 'x' : '') + '.png');
 
     return function(t) {
-        loadMap({ style: style, scale: scale }, function(err, map) {
+        new GL({ style: style }, function(err, map) {
             t.error(err);
 
             var callback = function(err, image) {
